transferred image change functionality to new PhotosDiv.js file
